backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
joblib
pydantic
python-multipart
httpx
python-dotenv
